feat(authentication): adds authentication
This commit adds authentication functionality to the app.
It adds the ability to login and register into the application.

It also adds the ability to reset your password. This is all
done by the devise library for ruby.

Functinality for admin to be able to view all users, show a
single user and delete users has also been added.

It also adds tests for all the above functionality.

[Finishes #159695478, #159695091, #159719400]