feat(vehicle): add vehicle functionality
This commit adds functionality to perform operations
on vehicles for both the admin and the user. An admin
can view, add, delete and update all vehicles added
to the application.

A user can view their own vehicles(vehicles they
added to the application), add, delete or update the
same.

It also adds tests for these functionalities.

[Finishes #159695949, #159696422, #159696368, #159696631]