feat(navbar): adds navigation bars
This commit adds navigation bars for the application. These are
to be used when navigating by the user to ensure easy navigation.
It does this by adding a header navigation which contains the links to
find a ride and add a ride, and a side bar which contains links
for crud operations.

It also adds structure to the application by providing it with a
wireframe for the structure, and giving it space for the app
content withing the app_container div.

[Finishes #159759994]